+++ plugins/parcel_css.ts
@@ -1,9 +1,0 @@
-import lightningcss from "./lightningcss.ts";
-
-console.log(
-  "%cparcel_css plugin was renamed to lightningcss. Please update your code.",
-  "color: red",
-);
-
-export * from "./lightningcss.ts";
-export default lightningcss;